// bookingBackend/src/interface/ticket.interface.ts
export interface Ticket {
    bookingId: string;
    busId: string;
    userName: string;
    email: string;
    mobileNumber: string;
    age: number;
    fromStation: string;
    toStation: string;
    seatNumber: number;
    isSingleLady: boolean;
    passengerType: string;
    fare: number;
    paymentType: string;
    pnrNumber: string;
  }
  